fix(types): make the mypy gate honest; fix bugs it caught (P0.5)
- Delete mypy.ini; pyproject.toml [tool.mypy] is the single config used
  by both bare runs and pre-commit/CI.
- Remove ignore_errors everywhere. Every module now checked with
  check_untyped_defs=true; unused-coroutine, unused-awaitable, call-arg,
  no-redef, name-defined, unused-ignore and return stay live across the
  whole tree (verified the gate fails when an await is removed).
- Bugs the honest gate caught, fixed here:
  * attachment_parser called claude_structured(tool_schema=...) — stale
    kwarg raised TypeError on every call, silently disabling AI column
    detection (regression test added).
  * safe_background_task held no strong reference to its Task while 17
    call sites discarded the return — the P0.4 GC bug at scale; the
    helper now retains tasks until done (test added).
- Remaining debt is explicit: per-module disable_error_code lists in
  pyproject.toml (~3k errors, mostly untyped Column() models — burn down
  via Mapped[] migration), recorded with counts in the plan doc.

Full suite: 22,712 passed. CI-path mypy: 0 issues in 533 files.

+++ app/company_utils.py
@@ -42,7 +42,7 @@ def _auto_keep_rank(company: dict) -> tuple[int, int, int, int]:
     )
 
 
-def _pair_dict(a: dict, b: dict, score: int) -> dict:
+def _pair_dict(a: dict, b: dict, score: float) -> dict:
     """Build one nested candidate dict (the public shape both backends emit).
 
     auto_keep_id follows the heuristic: more sites → has owner → is strategic → older id.
